% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player_function.R
\name{oldest_player_function}
\alias{oldest_player_function}
\title{NBA oldest_player_function Created by Hanzhang}
\usage{
oldest_player_function(year)
}
\arguments{
\item{oldest}{Do you want to know who  who is the oldest player in the NBA data for a chosen input year?}
}
\description{
This function allows you to know the player who is the oldest player in the NBA data for a chosen input year
}
\examples{
oldest_player_function()
}
\keyword{NBA,}
\keyword{Oldest}
